Editor's Manual

+++ Assets/Scripts/Data/Chart.cs
@@ -14,7 +14,7 @@ public class Chart : IStoryboardable, IDeepClonable<Chart>
     public string DifficultyName = "Normal";
     public string DifficultyLevel = "6";
     public int DifficultyIndex = 1;
-    public int ChartConstant = 6;
+    public float ChartConstant = 6;
 
     public List<LaneGroup> Groups = new List<LaneGroup>();
     public List<Lane> Lanes = new List<Lane>();
